requirements.txt
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
  frozenlist==1.5.0
asyncpg==0.30.0
databases==0.9.0
  SQLAlchemy==2.0.36
    greenlet==3.1.1
    typing_extensions==4.12.2
DateTime==5.5
  pytz==2024.2
  zope.interface==7.1.1
    setuptools==75.3.0
fastapi==0.115.4
  pydantic==2.9.2
    annotated-types==0.7.0
    pydantic_core==2.23.4
      typing_extensions==4.12.2
    typing_extensions==4.12.2
  starlette==0.41.2
    anyio==4.6.2.post1
      idna==3.10
      sniffio==1.3.1
  typing_extensions==4.12.2
orjson==3.10.11
pathlib==1.0.1
pytest==8.3.3
  colorama==0.4.6
  iniconfig==2.0.0
  packaging==24.1
  pluggy==1.5.0
python-dotenv==1.0.1
qrcode==8.0
  colorama==0.4.6
streamlit==1.40.0
  altair==5.4.1
    Jinja2==3.1.4
      MarkupSafe==3.0.2
    jsonschema==4.23.0
      attrs==24.2.0
      jsonschema-specifications==2024.10.1
        referencing==0.35.1
          attrs==24.2.0
          rpds-py==0.21.0
      referencing==0.35.1
        attrs==24.2.0
        rpds-py==0.21.0
      rpds-py==0.21.0
    narwhals==1.13.2
    packaging==24.1
    typing_extensions==4.12.2
  blinker==1.8.2
  cachetools==5.5.0
  click==8.1.7
    colorama==0.4.6
  GitPython==3.1.43
    gitdb==4.0.11
      smmap==5.0.1
  numpy==2.1.3
  packaging==24.1
  pandas==2.2.3
    numpy==2.1.3
    python-dateutil==2.9.0.post0
      six==1.16.0
    pytz==2024.2
    tzdata==2024.2
  pillow==11.0.0
  protobuf==5.28.3
  pyarrow==18.0.0
  pydeck==0.9.1
    Jinja2==3.1.4
      MarkupSafe==3.0.2
    numpy==2.1.3
  requests==2.32.3
    certifi==2024.8.30
    charset-normalizer==3.4.0
    idna==3.10
    urllib3==2.2.3
  rich==13.9.4
    markdown-it-py==3.0.0
      mdurl==0.1.2
    Pygments==2.18.0
  tenacity==9.0.0
  toml==0.10.2
  tornado==6.4.1
  typing_extensions==4.12.2
  watchdog==5.0.3
typing==3.7.4.3
uvicorn==0.32.0
  click==8.1.7
    colorama==0.4.6
  h11==0.14.0
validators==0.34.0
yarl==1.17.1
  idna==3.10
  multidict==6.1.0
  propcache==0.2.0
